requirements.txt
google-cloud-aiplatform>=1.55.0
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
importlib-metadata>=8.0.0
